pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "variant-viewer"
version = "0.1.0"
description = "Interactive variant effect viewer for Evo2 ClinVar probes"
requires-python = ">=3.12"

dependencies = [
    "polars~=1.36",
    "torch>=2.10,<2.11",
    "orjson",
    "numpy",
    "scikit-learn>=1.8",
    "umap-learn",
    "loguru>=0.7",
    "rich>=13",
    "typer>=0.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "goodfire-core>=0.3.0",
    "scipy>=1.14",
    "wandb>=0.23",
    "starlette>=0.45",
    "uvicorn[standard]",
    "duckdb>=1.2",
    "pyarrow>=23.0.1",
    "anthropic",
]

[project.scripts]
vv = "cli:app"

[dependency-groups]
dev = ["pytest>=9.0", "ruff>=0.15"]

[project.optional-dependencies]
train = ["wandb>=0.23"]

[tool.setuptools]
packages = ["probe", "pipeline"]
py-modules = ["cli", "build", "serve", "db", "transform", "constants", "display", "paths", "loaders", "prompts", "attribution", "training"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = [
    "B008",   # typer.Option() in defaults is the standard typer pattern
    "S108",   # /tmp for wandb SIGBUS avoidance is intentional
]

[tool.uv.sources]
goodfire-core = { git = "https://github.com/goodfire-ai/goodfire-core.git", tag = "v0.3.1" }
